requirements.txt
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
distro==1.9.0
Flask==3.1.3
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.8
